Added PartiesActivity.java, which retrieves the list of political parties from the database, sets the applications list of political parties (using the AsyncTask of the class GetPoliticalParties) and draws it on the screen inside a a GridView, with shows the custom view described in partyWidgetView.
GetPoliticalParties now also retrieves from the database the party logo (coded as base64 by the ServiceRest) and decodes it in order to store it as a Bitmap (which will be show in the GridView of the PartiesActivity). Also it shows a ProgressDialog in order to block the MainUI while downloading data.

PoliticalParty class now stores the logo as a Bitmap.

Added PoliticalProgramIndexActivity, which retrieves the political program of a given political party from the database and stores it on the corresponding application political party (using the AsyncTask of the class GetProgramsData).

GetProgramsData now shows a ProgressDialog in order to block the MainUI while downloading data. It also gets the id of the PoliticalParty whose program is being downloaded.