Added ExpandablIndexAdapter as an adapter to handle the Index of sections (2 levels) and properly show it on the View (PoliticalProgramIndexActivity)
If the data for the sections hasn't been downloaded yet, GetProgramsData downloads it and when it finishes it sets the structure for the index in order to set the adapter for the ExpandableListView. Otherwise (if the data has alreay been  previously downloaded) PoliticalProgramIndexActivity orders the index and sets the adapter.