Modified SectionViewerActivity: now shows the information of subsections in a simple list view (just only the next sublevel of sections) and the information of num likes, etc
PoliticalProgramIndex now shows only the main principal Sections (First level) in a simple List View (which uses ExpandableListAdapter). Every list item has its own listener for launching a new SectionViewerActivity with the info of the selected Section.

GetSectionContent now gets the information of a given Section from the database using a Get request, and storing all this data in the right section of the list of sections stored in the app.

Section now also stores the number of likes, dislikes, not undestood and comments.

PoliticalGroups now has a method that looks for the asked section of a political party program (looks into the tree structure)  and returns it.