Added class GetComments which implements the AsyncTask that gets the list of comments for a given Section and shows it to the user.
Added class Comment to store each comment of the list of comments.

Added CommentListAdapter to adapt the list of comments to the layout view.

Added comment_item layout to describe the layout (view) of a single comment.

Modified activity_comments layout to properly show the list of comments and the space for sending a new comment.