Modified ProgramsFragment so now it downloads (using GetTopIndex AsyncTask) and shows the top 3 ranking of sections ordered by different social elements (views, likes, comments, etc)
Modified Section so now it has new attributes (number of views) and a another new class constructor for this purpose.

Modified GetTopIndex so now it processes the JSON and stores the different sections on different lists depending on the social element that its been processed (likes, comments, views, dislikes, etc)

Modified fragment_programs layout so now it contains five listViews, one for each kind of social element top 3.

Added top_ranking_item layout that describes the layout for the view of an entry of a top ranking list.

Added TopItemAdapter to set the content of a listView for a given list of Sections.

Added